Code cleanup and bugproofing.
Renamed snapshotState to hasSnapshot for clarity.
Cleaned up and de-duplicated snapshot validation code, now contained in checkInvalidSnapshot(). Returns true when invalid.
Added inventory size check to onLoad() to match other validity tests.

SMP server will need to handle the case where the target inventory is in an unloaded chunk. Can we test if the server has the chunk containing a given block location loaded, and if not, simply bail out of all operations that test or modify the snapshot state, do nothing on redstone pulse, etc? in order to prevent odd behavior such as improperly clearing the snapshot because it gets a null tile entity due to an unloaded chunk...