Stocker now detects adjacent RedPower tubes and BuildCraft pipes, and 'opens' access ports.
Limitations:

Cannot detect if a cover is preventing a tube from connecting to the Stocker.

Cannot currently differentiate waterproof and conductive pipes from item transport pipes.

Either situation will result in an erroneously 'open' texture. Functionality is unaffected; this is a display issue only, but may prove to be unavoidable.