Fix NPE after saving and reloading when stocker has snapshot of double chest. Read from NBT was not creating second array. Don't know why that wasn't throwing tantrums, I mean exceptions, of its own.
Revert "Revert "Attempt at improving visual of redstone connections to Stocker.""

This reverts commit 5ac8ef633c5137150a51cbaebc979833917aec27.

Adding this back in because I was actually doing it right and RedPower was doing it wrong until PR5b2, lawl.
It still causes redstone dust to appear to connect in some cases when it really isn't, but that's not a problem on our end.